fix(frontend): banner constants moved
