+++ packages/frontend/constants/index.ts
@@ -2,4 +2,3 @@ export * from './asset';
 export * from './common';
 export * from './notifications';
 export * from './ui';
-export * from './banners';
